% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{as_ROH}
\alias{as_ROH}
\title{Deprecated: use \code{\link{as_RUNS}} instead}
\usage{
as_ROH(...)
}
\arguments{
\item{...}{passed to \code{as_RUNS}}
}
\description{
Deprecated: use \code{\link{as_RUNS}} instead
}
\keyword{internal}
