% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chillingham_genotype}
\alias{chillingham_genotype}
\title{chillingham_genotype}
\format{A data frame with 17 rows and 2641 variables}
\source{
Chillingham dataset (cite)
}
\usage{
chillingham_genotype
}
\description{
A dataset containing the 0/1/2 SNP genotypes of 17 Chillingham cattle on BTA1
}
\details{
\itemize{
  \item{\code{FID}}{}
  \item{\code{IID}}: id of animals
  \item{\code{PAT}}{}
  \item{\code{MAT}}{}
  \item{\code{SEX}}{}
  \item{\code{PHENOTYPE}}{}
  \item{\code{BovineHD0100015049_0}}: sample name
  \item{\code{BovineHD0100015050_0}}: sample name
  \item{\code{BovineHD0100015051_0}}: sample name
  \item{\code{...}}
}
}
\examples{
data(chillingam)
head(chillingham_genotype)

}
\keyword{datasets}

