% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chillingham_map}
\alias{chillingham_map}
\title{chillingham_map}
\format{A data frame with 2635 rows and 4 variables}
\source{
Chillingham dataset (cite)
}
\usage{
chillingham_map
}
\description{
A dataset containing the Map file for SNP on BTA1
}
\details{
\describe{
  \item{\code{Chrom}}{chromosome name}
  \item{\code{SNP}}{SNP name}
  \item{\code{cM}}{position in cM}
  \item{\code{bps}}{position in bp}
}
}
\examples{
data(chillingam)
head(chillingham_map)

}
\keyword{datasets}

