% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stats.R
\name{chromosomeLength}
\alias{chromosomeLength}
\title{Function to found max position for each chromosome}
\usage{
chromosomeLength(mapFile)
}
\arguments{
\item{mapFile}{Plink map file (for SNP position)}
}
\value{
A data frame with the max position for chromosome
}
\description{
Function to found max position for each chromosome
}
\details{
Create a data frame with the max position in map file (plink format)
}
\keyword{internal}
