% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{consecutiveRuns}
\alias{consecutiveRuns}
\title{Function to detect consecutive runs in a vector (individual's genotypes)}
\usage{
consecutiveRuns(indGeno, individual, mapFile, ROHet = TRUE, minSNP = 3,
  maxOppositeGenotype = 1, maxMiss = 1, minLengthBps = 1000,
  maxGap = 10^6)
}
\arguments{
\item{indGeno}{vector of 0/1/NAs of individual genotypes (0: homozygote; 1: heterozygote)}

\item{individual}{group (breed, population, case/control etc.) and ID of individual sample}

\item{mapFile}{Plink map file (for SNP position)}

\item{ROHet}{shall we detect ROHet or ROHom?}

\item{minSNP}{minimum number of SNP in a run}

\item{maxOppositeGenotype}{max n. of homozygous/heterozygous SNP}

\item{maxMiss}{max. n. of missing SNP}

\item{minLengthBps}{min length of a run in bps}

\item{maxGap}{max distance between consecutive SNP in a window to be stil considered a potential run}
}
\value{
data frame of runs per individual
}
\description{
This is a core function. It implements the consecutive method for run detection
}
\examples{

}

