% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{consecutiveRUNS.run}
\alias{consecutiveRUNS.run}
\title{Main function to detect genomic RUNS (ROHom/ROHet) using the consecutive method}
\usage{
consecutiveRUNS.run(
  genotypeFile,
  mapFile,
  ROHet = FALSE,
  maxOppRun = 0,
  maxMissRun = 0,
  minSNP = 15,
  minLengthBps = 1000,
  maxGap = 10^6,
  nCores = NULL
)
}
\arguments{
\item{genotypeFile}{genotype (.ped) file path}

\item{mapFile}{map file (.map) file path}

\item{ROHet}{should we look for ROHet or ROHom? (default = FALSE)}

\item{maxOppRun}{max n. of opposite genotype SNPs in the run (default = 0)}

\item{maxMissRun}{max n. of missing SNPs in the run (default = 0)}

\item{minSNP}{minimum n. of SNP in a RUN (default = 15)}

\item{minLengthBps}{minimum length of run in bps (defaults to 1000 bps = 1 kbps)}

\item{maxGap}{max distance between consecutive SNP in a window to be still
considered a potential run (defaults to 10^6)}

\item{nCores}{number of cores to use (deprecated, ignored)}
}
\value{
A dataframe with RUNs of Homozygosity or Heterozygosity.
}
\description{
\strong{Deprecated.}
}
\details{
This function is deprecated. Use \code{\link{scanRUNS}} instead.
}
\examples{
# getting map and ped paths
genotypeFile <- system.file("extdata", "Kijas2016_Sheep_subset.ped", package = "detectRUNS")
mapFile <- system.file("extdata", "Kijas2016_Sheep_subset.map", package = "detectRUNS")
\dontrun{
runs <- consecutiveRUNS.run(genotypeFile, mapFile, minSNP = 15, ROHet = FALSE,
maxOppRun = 0, maxMissRun = 0, maxGap=10^6, minLengthBps = 100000)
}
runsFile <- system.file("extdata", "Kijas2016_Sheep_subset.consecutive.csv", package="detectRUNS")
colClasses <- c(rep("character", 3), rep("numeric", 4))
runs <- read.csv2(runsFile, header = TRUE, stringsAsFactors = FALSE, colClasses = colClasses)

}
