% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{countSnpInRunsCpp}
\alias{countSnpInRunsCpp}
\title{Function to count number of times a SNP is in a RUN}
\usage{
countSnpInRunsCpp(runsChrom, mapChrom, genotypeFile, threshold = 50)
}
\arguments{
\item{runsChrom}{R object (dataframe) with results per chromosome}

\item{mapChrom}{R map object with SNP per chromosome}

\item{genotypeFile}{genotype (.ped) file location}

\item{threshold}{filter out SNPs below this value}
}
\value{
dataframe with counts per SNP in runs (per population)
}
\description{
Similar to snpInsideRunsCpp, it apply additional column to dataframe.
Require to ensure backward compatibility.
}
