% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNS.R
\name{.get_chrom_lengths}
\alias{.get_chrom_lengths}
\title{Get chromosome lengths from a RUNS object or by reading mapFile}
\usage{
.get_chrom_lengths(x, mapFile = NULL)
}
\description{
Get chromosome lengths from a RUNS object or by reading mapFile
}
\keyword{internal}
