% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROH.R
\name{.get_runs}
\alias{.get_runs}
\title{Extract the runs data.frame from an ROH object or return as-is}
\usage{
.get_runs(x)
}
\description{
Extract the runs data.frame from an ROH object or return as-is
}
\keyword{internal}
