% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROH.R
\name{.get_sample_info}
\alias{.get_sample_info}
\title{Get sample info from an ROH object or by reading genotypeFile}
\usage{
.get_sample_info(x, genotypeFile = NULL)
}
\description{
Get sample info from an ROH object or by reading genotypeFile
}
\keyword{internal}
