% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{findOppositeAndMissing}
\alias{findOppositeAndMissing}
\title{Function to calculate oppositeAndMissingGenotypes array}
\usage{
findOppositeAndMissing(data, ROHet = TRUE)
}
\arguments{
\item{data}{vector of 0/1/2 genotypes}

\item{ROHet}{TRUE in ROHet evaluation, FALSE for ROHom}
}
\value{
character array; names will be index in which opposite and missing
snps are found in data array
}
\description{
This is an helper function, this will be called by another function
}
