% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{genoUmschalten}
\alias{genoUmschalten}
\title{Function to convert 0/1/2 genotypes to 0/1 (either homozygous/heterozygous)}
\usage{
genoUmschalten(x)
}
\arguments{
\item{x}{vector of 0/1/2 genotypes}
}
\value{
converted vector of genotypes
}
\description{
This is a utility function.
}
\examples{
#not yet


}

