% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{heteroZygotPrufen}
\alias{heteroZygotPrufen}
\title{Function to check whether a window is (loosely) heterozygous or not}
\usage{
heteroZygotPrufen(x, gaps, maxHom, maxMiss, maxGap)
}
\arguments{
\item{x}{vector of 0/1 genotypes (from genoUmschalten())}

\item{maxHom}{max n. of homozygous SNP in a heterozygous window}

\item{maxMiss}{max n. of missing in a window}
}
\value{
TRUE/FALSE (whether a window is heterozygous or NOT)
}
\description{
This is a core function. Parameters on how to consider a window heterozygous are here (maxHom, maxMiss)
}
\examples{
#not yet


}

