% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{heteroZygotTest}
\alias{heteroZygotTest}
\title{Function to check whether a window is (loosely) heterozygous or not}
\usage{
heteroZygotTest(x, gaps, maxHom, maxMiss, maxGap, i, windowSize)
}
\arguments{
\item{x}{vector of 0/1 genotypes (from genoConvert())}

\item{gaps}{vector of differences between consecutive positions (gaps) in bps}

\item{maxHom}{max n. of homozygous SNP in a heterozygous window}

\item{maxMiss}{max n. of missing in a window}

\item{maxGap}{max distance between consecutive SNP in a window to be still considered a potential run}

\item{i}{index along the genome (genome-vector for each individual)}

\item{windowSize}{size of window (n. of SNP)}
}
\value{
a list: i) TRUE/FALSE (whether a window is heterozygous or NOT); ii) indexes of "opposite and missing" genotype
}
\description{
This is a core function within the sliding-window workflow. Parameters on how to consider a window heterozygous are here (maxHom, maxMiss)
}
\examples{
\dontrun{
maxHom <- 1
maxMiss <- 1
maxGap <- 10^6
i <- 150 #random index
x <- c(0, 0, NA, 0, NA, 0, 0, 0, 1, 1,
       1, 1, 1, 1, 0, 0, 1, 0, 0, 0)
gaps <- c(4374, 8744, 5123, 14229, 5344, 690, 8566, 5853, 2369, 3638,
          4848, 600, 2333, 976, 2466, 2269, 5411, 6021, 4367)
windowSize <- length(x)
test <- heteroZygotTest(x, gaps, maxHom, maxMiss, maxGap,i,windowSize)
}
# test
# $windowStatus
# [1] FALSE
#
# $oppositeAndMissingSNP
# 150 151 152 153 154 155 156 157 164 165 167 168 169
#   0   0   9   0   9   0   0   0   0   0   0   0   0
}
