% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{heteroZygotTestCpp}
\alias{heteroZygotTestCpp}
\title{Function to check whether a window is (loosely) heterozygous or not}
\usage{
heteroZygotTestCpp(x, gaps, maxHom, maxMiss, maxGap)
}
\arguments{
\item{x}{vector of 0/1 genotypes (from genoConvert())}

\item{gaps}{vector of differences between consecutive positions (gaps) in bps}

\item{maxHom}{max n. of homozygous SNP in a heterozygous window}

\item{maxMiss}{max n. of missing in a window}

\item{maxGap}{max distance between consecutive SNP in a window to be stil considered a potential run}
}
\value{
TRUE/FALSE (whether a window is heterozygous or NOT)
}
\description{
This is a core function. Parameters on how to consider a window heterozygous are here (maxHom, maxMiss)
}
\examples{
maxHom <- 1
maxMiss <- 1
maxGap <- 10^6
x <- c(0, 0, NA, 0, NA, 0, 0, 0, 1, 1,
       1, 1, 1, 1, 0, 0, 1, 0, 0, 0)
gaps <- c(4374, 8744, 5123, 14229, 5344, 690, 8566, 5853, 2369, 3638,
          4848, 600, 2333, 976, 2466, 2269, 5411, 6021, 4367)
test <- heteroZygotTestCpp(x, gaps, maxHom, maxMiss, maxGap)
# test is FALSE

}

