% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{homoZygotTest}
\alias{homoZygotTest}
\title{Function to check whether a window is (loosely) homozygous or not}
\usage{
homoZygotTest(x, gaps, maxHet, maxMiss, maxGap, i, windowSize)
}
\arguments{
\item{x}{vector of 0/1 genotypes (from genoConvert())}

\item{gaps}{vector of differences between consecutive positions (gaps) in bps}

\item{maxHet}{max n. of heterozygous SNP in a homozygous window}

\item{maxMiss}{max n. of missing in a window}

\item{maxGap}{max distance between consecutive SNP in a window to be stil considered a potential run}

\item{i}{index along the genome (genome-vector for each individual)}

\item{windowSize}{size of window (n. of SNP)}
}
\value{
TRUE/FALSE (whether a window is homozygous or NOT)
}
\description{
This is a core function. Parameters on how to consider a window homozygous are here (maxHet, maxMiss)
}
\examples{
maxHom <- 1
maxMiss <- 1
maxGap <- 10^6
i <- 175
x <- c(0, 0, 0, NA, 0, 0, 1, 0, 0, 0,
       0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
gaps <- c(3721, 3871, 7059, 4486, 7545, 4796, 3043, 9736, 3495, 5051,
          9607, 6555, 11934, 6410, 3415, 1302, 3110, 6609, 3292)
windowSize <- length(x)
test <- homoZygotTest(x, gaps, maxHom, maxMiss, maxGap,i,windowSize)
# test is true
x <- c(0, 1, 1, 1, 1, 1, 1, 1, 1, 1,
       1, 1, 1, 1, 1, 1, 1, 1, 1, 1)
gaps <- c(2514, 2408, 2776, 2936, 1657, 494, 1436, 680, 909, 678,
          615, 1619, 2058, 2446, 1085, 660, 1259, 1042, 2135)
test <- homoZygotTest(x, gaps, maxHom, maxMiss, maxGap)
# test is false

}

