% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{loadROH}
\alias{loadROH}
\title{Deprecated: use \code{\link{loadRUNS}} instead}
\usage{
loadROH(...)
}
\arguments{
\item{...}{passed to \code{loadRUNS}}
}
\description{
Deprecated: use \code{\link{loadRUNS}} instead
}
\keyword{internal}
