% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{manhattan_Runs}
\alias{manhattan_Runs}
\title{Plot N. of times SNP is in runs - MANHATTAN PLOT}
\usage{
manhattan_Runs(runs, genotype_path, mapfile_path, savePlots = FALSE,
  title_prefix = NULL, main_titel = NULL)
}
\arguments{
\item{runs}{a data.frame with runs per animal (breed, id, chrom, nSNP, start, end, length)}

\item{genotype_path}{genotype (.ped) file location}

\item{mapfile_path}{map file (.map) file location}

\item{savePlots}{should plots be saved out in files (default) or plotted in the graphical terminal?}

\item{title_prefix}{title prefix (the base name of graph, if savePlots is TRUE)}

\item{main_titel}{title in plot #FILIPPO}
}
\value{
plot of n. of times a SNP is in a run by chromosome and population (pdf files) using manhattan
}
\description{
Function to plot the number of times/percentage a SNP in in a run (population-specific signals)
Proportions on the y-axis, bps on the x-axis
}
\examples{
# getting map and ped paths
genotype_path <- system.file("extdata", "subsetChillingham.ped", package = "detectRUNS")
mapfile_path <- system.file("extdata", "subsetChillingham.map", package = "detectRUNS")

# calculating runs of Homozygosity
runs <- RUNS.run(genotype_path, mapfile_path, windowSize = 20, threshold = 0.1, minSNP = 5,
ROHet = FALSE, maxOppositeGenotype = 1, maxMiss = 1,  minLengthBps = 1000, minDensity = 1/10)

# plot runs per animal (interactive)
manhattan_Runs(runs, genotype_path, mapfile_path, savePlots=FALSE, title_prefix="ROHom")

}

