% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{pedConvertCpp}
\alias{pedConvertCpp}
\title{Convert ped genotypes to 0/1}
\usage{
pedConvertCpp(genotype)
}
\arguments{
\item{genotype}{vector of pair of genotypes (01, AA, AG)}
}
\value{
converted vector of genotypes (0/1)
}
\description{
This is a utility function, that convert ped genotypes (AA/AB/BB) into 0/1
(either homozygous/heterozygous)
}
