% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStackedRuns}
\alias{plotStackedRuns}
\title{Plot stacked runs}
\usage{
plotStackedRuns(runs, savePlots = FALSE, title_prefix = NULL)
}
\arguments{
\item{runs}{a data.frame with runs per animal (breed, id, chrom, nSNP, start, end, length)}

\item{savePlots}{should plots be saved out in files (default) or plotted in the graphical terminal?}

\item{title_prefix}{title prefix (the base name of graph, if savePlots is TRUE)}
}
\value{
plot of stacked runs by population and by chromosome (pdf files)
}
\description{
Function to plot stacked runs along the chromosome (signalling presence of large numbers of runs)
Counts on the y-axis, bps on the x-axis: plots run (TRUE) / no run (FALSE)
}
\examples{
# getting map and ped paths
genotype_path <- system.file("extdata", "subsetChillingham.ped", package = "detectRUNS")
mapfile_path <- system.file("extdata", "subsetChillingham.map", package = "detectRUNS")

# calculating runs of Homozygosity
runs <- RUNS.run(genotype_path, mapfile_path, windowSize = 20, threshold = 0.1,
minSNP = 5, ROHet = FALSE, maxOppositeGenotype = 1, maxMiss = 1,
minLengthBps = 1000, minDensity = 1/10)

# plot runs per animal (interactive)
plotStackedRuns(runs, savePlots=FALSE, title_prefix="ROHom")

}

