% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROH.R
\name{print.ROH}
\alias{print.ROH}
\title{Print a summary of an ROH object}
\usage{
\method{print}{ROH}(x, ...)
}
\arguments{
\item{x}{An \code{ROH} object returned by \code{\link{scanRUNS}} or
\code{\link{as_ROH}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary of an ROH object
}
\examples{
\dontrun{
bedFile <- system.file("extdata", "Kijas2016_Sheep_subset.bed",
                        package = "detectRUNS")
roh <- scanRUNS(bedFile, method = "sliding", minSNP = 15,
                minLengthBps = 100000)
print(roh)
}
}
