% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rohIslands.R
\name{print.ROHIslands}
\alias{print.ROHIslands}
\title{Print a summary of an ROHIslands object}
\usage{
\method{print}{ROHIslands}(x, ...)
}
\arguments{
\item{x}{An \code{ROHIslands} object returned by \code{\link{rohIslands}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary of an ROHIslands object
}
