% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNS.R
\name{print.RUNS}
\alias{print.RUNS}
\title{Print a summary of a RUNS object}
\usage{
\method{print}{RUNS}(x, ...)
}
\arguments{
\item{x}{A \code{RUNS} object returned by \code{\link{scanRUNS}} or
\code{\link{as_RUNS}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary of a RUNS object
}
\examples{
\dontrun{
bedFile <- system.file("extdata", "Kijas2016_Sheep_subset.bed",
                        package = "detectRUNS")
roh <- scanRUNS(bedFile, method = "sliding", minSNP = 15,
                minLengthBps = 100000)
print(roh)
}
}
