% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runsIslands.R
\name{print.RunsIslands}
\alias{print.RunsIslands}
\title{Print a summary of an RunsIslands object}
\usage{
\method{print}{RunsIslands}(x, ...)
}
\arguments{
\item{x}{An \code{RunsIslands} object returned by \code{\link{runsIslands}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print a summary of an RunsIslands object
}
