% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{readFromPlink}
\alias{readFromPlink}
\title{READ ROH OUTPUT FILE FROM PLINK
Function to read in the output file from ROH analysis with Plink
Relevant columns are selected, converted and renamed}
\usage{
readFromPlink(plinkFile = "plink.hom")
}
\arguments{
\item{plinkFile}{name of output file from Plink ROH analysis #defaults to plink.hom}
}
\value{
data frame formatted to be used with plot and statistics functions (package detectRUNS)
}
\description{
READ ROH OUTPUT FILE FROM PLINK
Function to read in the output file from ROH analysis with Plink
Relevant columns are selected, converted and renamed
}
\examples{
#not yet


}

