% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{readMapFile}
\alias{readMapFile}
\title{Read from a .map file locations and return a data.table object}
\usage{
readMapFile(mapFile)
}
\arguments{
\item{mapFile}{map file (.map) file path}
}
\value{
data.table object
}
\description{
This is an utility function which check for file existance, define 
colClasses and then returns the read data.table object
}
\keyword{internal}
