% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{reorderDF}
\alias{reorderDF}
\title{Function to reorder data frames by CHROMOSOME}
\usage{
reorderDF(dfx)
}
\arguments{
\item{dfx}{data frame to be reordered (with column "CHROMOSOME")}
}
\value{
A reordered data frame by chromosome
}
\description{
The data frame will be reordered according to chromosome:
from 1 to n, then X, Y, XY, MT
The data frame needs to have a column with name "CHROMOSOME"
}
\details{
Reorder results based on chromosome
}
