% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{RUNS.run}
\alias{RUNS.run}
\title{Run functions to detect RUNS (ROHom/ROHet)}
\usage{
RUNS.run(gegevens, mapFile, windowSize = 15, drempel = 0.1, minSNP = 3,
  ROHet = TRUE, maxOppositeGenotype = 1, maxMiss = 1, maxGap = 10^6,
  minLengthBps = 1000, minDensity = 1/10)
}
\arguments{
\item{gegevens}{genotype dataset in the Plink .raw format}

\item{drempel}{the threshold of overlapping windows of the same state (homozygous/heterozygous) to call a SNP in a RUN}

\item{minSNP}{minimum n. of SNP in a RUN}

\item{ROHet}{should we look for ROHet or ROHom?}

\item{maxGap}{max distance between consecutive SNP in a window to be stil considered a potential run}

\item{minLengthBps}{minimum length of run in bps (defaults to 1000 bps = 1 kbps)}

\item{minDensity}{minimum n. of SNP per kbps (defaults to 0.1 = 1 SNP every 10 kbps)}

\item{mappa}{Plink map file (either the file path/name or the R data.frame)}

\item{window}{the size of sliding window}
}
\value{
n. of individuals for which runs have been written out
}
\description{
This is the main function of RUNS and would probably require
some good documentation.
}
\examples{
x <- RUNS.run(gegevens, mappa, windowSize = 20, drempel = 0.1, minSNP = 5,
ROHet = TRUE, maxOppositeGenotype = 1, maxMiss = 1,  minLengthBps = 1000, minDensity = 1/10)

}

