% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{schreibRUN}
\alias{schreibRUN}
\title{Function to write out RUNS per individual animal}
\usage{
schreibRUN(ind, dRUN)
}
\arguments{
\item{ind}{ID of animal}

\item{dRUN}{data.frame with RUNS per animal}
}
\value{
TRUE/FALSE if RUNS are written out or not
}
\description{
Function to write out RUNS per individual animal
}
\examples{
#not yet


}

