% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{snpInsideRuns}
\alias{snpInsideRuns}
\title{Function to count number of times a SNP is in a RUN. Need to be called per
chromosome (not using a dataframe with all results on all chromosomes)}
\usage{
snpInsideRuns(runsChrom, mapChrom, sample_info)
}
\arguments{
\item{runsChrom}{R object (dataframe) with results per chromosome (column names:"POPULATION","IND","CHROMOSOME","COUNT","START","END","LENGTH")}

\item{mapChrom}{R object (dataframe) with SNP name and position per chromosome (map file) (column names: "CHR","SNP_NAME","POSITION")}

\item{sample_info}{data.frame with columns group and id (one row per individual)}
}
\value{
dataframe with counts per SNP in runs (per population)
}
\description{
Function to count number of times a SNP is in a RUN. Need to be called per
chromosome (not using a dataframe with all results on all chromosomes)
}
\keyword{internal}
