% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{tableRunsCpp}
\alias{tableRunsCpp}
\title{Function to retrieve most common runs in the population}
\usage{
tableRunsCpp(runs, genotypeFile, mapFile, threshold = 0.5)
}
\arguments{
\item{runs}{R object (dataframe) with results on detected runs}

\item{genotypeFile}{Plink ped file (for SNP position)}

\item{mapFile}{Plink map file (for SNP position)}

\item{threshold}{value from 0 to 1 (default 0.5) that controls the desired
proportion of individuals carrying that run (e.g. 50\%)}
}
\value{
A dataframe with the most common runs detected in the sampled individuals
(the group/population, start and end position of the run, chromosome, number of SNP
included in the run and average percentage of SNPs in run
are reported in the output dataframe)
}
\description{
This function takes in input either the run results and returns a subset of
the runs most commonly found in the group/population. The parameter
\code{threshold} controls the definition
of most common (e.g. in at least 50\%, 70\% etc. of the sampled individuals)
}
\examples{
# getting map and ped paths
genotypeFile <- system.file("extdata", "Kijas2016_Sheep_subset.ped", package = "detectRUNS")
mapFile <- system.file("extdata", "Kijas2016_Sheep_subset.map", package = "detectRUNS")

# calculating runs of Homozygosity
\dontrun{
# skipping runs calculation
runs <- scanRUNS(genotypeFile, method = "sliding", windowSize = 15, threshold = 0.1, minSNP = 15,
  ROHet = FALSE, maxOpp = 1, maxMiss = 1, minLengthBps = 100000
)
}
# loading pre-calculated data
runsFile <- system.file("extdata", "Kijas2016_Sheep_subset.sliding.csv", package = "detectRUNS")
runsDF <- readExternalRuns(inputFile = runsFile, program = "detectRUNS")

table <- tableRunsCpp(
  runs = runsDF, genotypeFile = genotypeFile,
  mapFile = mapFile, threshold = 0.5)

}
