% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{writeRUN}
\alias{writeRUN}
\title{Function to write out RUNS per individual animal}
\usage{
writeRUN(ind, dRUN, ROHet = TRUE, group)
}
\arguments{
\item{ind}{ID of animal}

\item{dRUN}{data.frame with RUNS per animal}

\item{ROHet}{shall we detect ROHet or ROHom?}

\item{group}{group (factor): population, breed, ethnicity, case/control etc.}
}
\value{
TRUE/FALSE if RUNS are written out or not
}
\description{
Function to write out RUNS per individual animal
}
\examples{
#not yet

}
