% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{classifyRuns}
\alias{classifyRuns}
\title{Classify runs in bins.}
\usage{
classifyRuns(runs, class_size = 2)
}
\arguments{
\item{runs}{a ROH dataframe object}

\item{class_size}{the starting bin size}
}
\value{
a new dataframe with two new columns, MB for ROH length in MB and
CLASS column which tag a ROH in a proper bin relying on size
}
\description{
Classify runs in bins.
}
