% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ROH.R
\name{.get_snp_map}
\alias{.get_snp_map}
\title{Get SNP map from an ROH object or by reading mapFile}
\usage{
.get_snp_map(x, mapFile = NULL)
}
\description{
Get SNP map from an ROH object or by reading mapFile
}
\keyword{internal}
