% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{heteroZygotTestCpp}
\alias{heteroZygotTestCpp}
\title{Function to check whether a window is (loosely) heterozygous or not}
\usage{
heteroZygotTestCpp(x, gaps, maxHom, maxMiss, maxGap)
}
\arguments{
\item{x}{vector of 0/1 genotypes (from genoConvert())}

\item{gaps}{vector of differences between consecutive positions (gaps) in bps}

\item{maxHom}{max n. of homozygous SNP in a heterozygous window}

\item{maxMiss}{max n. of missing in a window}

\item{maxGap}{max distance between consecutive SNP in a window to be still considered a potential run}
}
\value{
TRUE/FALSE (whether a window is heterozygous or NOT)
}
\description{
This is a core function. Parameters on how to consider a window heterozygous are here (maxHom, maxMiss)
}
