% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{homoZygotTest}
\alias{homoZygotTest}
\title{Function to check whether a window is (loosely) homozygous or not}
\usage{
homoZygotTest(x, gaps, maxHet, maxMiss, maxGap, i, windowSize)
}
\arguments{
\item{x}{vector of 0/1 genotypes (from genoConvert())}

\item{gaps}{vector of differences between consecutive positions (gaps) in bps}

\item{maxHet}{max n. of heterozygous SNP in a homozygous window}

\item{maxMiss}{max n. of missing in a window}

\item{maxGap}{max distance between consecutive SNP in a window to be stil considered a potential run}

\item{i}{index along the genome (genome-vector for each individual)}

\item{windowSize}{size of window (n. of SNP)}
}
\value{
TRUE/FALSE (whether a window is homozygous or NOT)
}
\description{
This is a core function. Parameters on how to consider a window homozygous are here (maxHet, maxMiss)
}
\examples{
maxHet <- 1
maxMiss <- 1
maxGap <- 10^6
i <- 175
x <- c(0, 0, 0, NA, 0, 0, 1, 0, 0, 0,
       0, 0, 0, 0, 0, 0, 0, 0, 0, 0)
gaps <- c(3721, 3871, 7059, 4486, 7545, 4796, 3043, 9736, 3495, 5051,
          9607, 6555, 11934, 6410, 3415, 1302, 3110, 6609, 3292)
windowSize <- length(x)
test <- homoZygotTest(x, gaps, maxHet, maxMiss, maxGap, i, windowSize)
# test
# $windowStatus
# [1] TRUE
#
# $oppositeAndMissingSNP
# 178 181
#   9   0

}

