% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{homoZygotTestCpp}
\alias{homoZygotTestCpp}
\title{Function to check whether a window is (loosely) homozygous or not}
\usage{
homoZygotTestCpp(x, gaps, maxHet, maxMiss, maxGap)
}
\arguments{
\item{x}{vector of 0/1 genotypes (from genoConvert())}

\item{gaps}{vector of differences between consecutive positions (gaps) in bps}

\item{maxHet}{max n. of heterozygous SNP in a homozygous window}

\item{maxMiss}{max n. of missing in a window}

\item{maxGap}{max distance between consecutive SNP in a window to be still considered a potential run}
}
\value{
TRUE/FALSE (whether a window is homozygous or NOT)
}
\description{
This is a core function. Parameters on how to consider a window homozygous are here (maxHet, maxMiss)
}
