% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plot_SnpsInRuns}
\alias{plot_SnpsInRuns}
\title{Plot the number of times each SNP falls inside runs}
\usage{
plot_SnpsInRuns(
  runs,
  genotypeFile = NULL,
  mapFile = NULL,
  savePlots = FALSE,
  separatePlots = FALSE,
  outputName = NULL
)
}
\arguments{
\item{runs}{a data.frame with runs per individual (group, id, chrom, nSNP, start, end, length)}

\item{genotypeFile}{genotype (.ped) file path}

\item{mapFile}{map file (.map) file path}

\item{savePlots}{should plots be saved out in files (default) or plotted in
the graphical terminal?}

\item{separatePlots}{should plots for each chromosome be saved out to separate files?}

\item{outputName}{title prefix (the base name of graph, if savePlots is TRUE)}
}
\value{
plot number of times a SNP is in a run by chromosome and population (pdf files)
}
\description{
Function to plot the number of times/percentage each SNP is inside a run
(population-specific signals) against the SNP positions in the genome.
Proportions on the y-axis, bps on the x-axis
}
\examples{
# getting map and ped paths
genotypeFile <- system.file("extdata", "Kijas2016_Sheep_subset.ped", package = "detectRUNS")
mapFile <- system.file("extdata", "Kijas2016_Sheep_subset.map", package = "detectRUNS")

# calculating runs of Homozygosity
# skipping runs calculation
\dontrun{
runs <- scanRUNS(genotypeFile, method = "sliding", windowSize = 15, threshold = 0.1, minSNP = 15,
ROHet = FALSE, maxOpp = 1, maxMiss = 1, minLengthBps = 100000)
}
# loading pre-calculated data
runsFile <- system.file("extdata", "Kijas2016_Sheep_subset.sliding.csv", package="detectRUNS")
runsDF <- readExternalRuns(inputFile = runsFile, program = 'detectRUNS')

# plot runs per animal (interactive)
plot_SnpsInRuns(runs = runsDF, genotypeFile = genotypeFile, mapFile = mapFile,
savePlots = FALSE, outputName = "ROHom")

}
