% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotRuns}
\alias{plotRuns}
\title{Function to plot runs per animal (see Williams et al. 2016, Animal Genetics)
IDs on the y-axis, bps on the x-axis: plots run (TRUE) / no run (FALSE)}
\usage{
plotRuns(runsFile = "detected.ROHet.csv", suppressInds = FALSE)
}
\arguments{
\item{suppressInds}{shall we suppress individual IDs on the y-axis? (defaults to FALSE)}

\item{runs}{output file with runs per animal (breed, id, chrom, nSNP, start, end, length) #defaults to detectRUNS.ROHet.csv}
}
\value{
plot of runs by chromosome (pdf files)
}
\description{
Function to plot runs per animal (see Williams et al. 2016, Animal Genetics)
IDs on the y-axis, bps on the x-axis: plots run (TRUE) / no run (FALSE)
}
\examples{
#not yet


}

