% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSnpsInRuns}
\alias{plotSnpsInRuns}
\title{PLOT N. of TIMES SNP IS IN RUNS
Function to plot the number of times/percentage a SNP in in a run (population-specific signals)
Proportions on the y-axis, bps on the x-axis}
\usage{
plotSnpsInRuns(runsFile = "detected.ROHet.csv", mapFile = "plink.map",
  rawFile = "genotype.raw", savePlots = TRUE)
}
\arguments{
\item{runsFile}{output file with runs per animal (breed, id, chrom, nSNP, start, end, length) #defaults to detectRUNS.ROHet.csv}

\item{mapFile}{map file (optional) #defaults to plink.map}

\item{rawFile}{raw file of 0/1/2 genotypes (optional) #defaults to genotype.raw}

\item{savePlots}{should plots be saved out in files (default) or plotted in the graphical terminal?}
}
\value{
plot of n. of times a SNP is in a run by chromosome and population (pdf files)
}
\description{
PLOT N. of TIMES SNP IS IN RUNS
Function to plot the number of times/percentage a SNP in in a run (population-specific signals)
Proportions on the y-axis, bps on the x-axis
}
\examples{
#not yet


}

