% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotSnpsInRuns}
\alias{plotSnpsInRuns}
\title{PLOT N. of TIMES SNP IS IN RUNS
Function to plot the number of times/percentage a SNP in in a run (population-specific signals)
Proportions on the y-axis, bps on the x-axis}
\usage{
plotSnpsInRuns(runsFile = "detected.ROHet.csv", mapFile = "karte.map",
  rawFile = "gegevens.raw")
}
\arguments{
\item{rawFile}{raw file of 0/1/2 genotypes (optional) #defaults to gegevens.raw}

\item{runs}{output file with runs per animal (breed, id, chrom, nSNP, start, end, length) #defaults to detectRUNS.ROHet.csv}

\item{mappa}{map file (optional) #defaults to karte.map}
}
\value{
plot of n. of times a SNP is in a run by chromosome and population (pdf files)
}
\description{
PLOT N. of TIMES SNP IS IN RUNS
Function to plot the number of times/percentage a SNP in in a run (population-specific signals)
Proportions on the y-axis, bps on the x-axis
}
\examples{
#not yet


}

