% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plots.R
\name{plotStackedRuns}
\alias{plotStackedRuns}
\title{PLOT STACKED RUNS
Function to plot stacked runs along the chromosome (signalling presence of large numbers of runs)
Counts on the y-axis, bps on the x-axis: plots run (TRUE) / no run (FALSE)}
\usage{
plotStackedRuns(runsFile = "detected.ROHet.csv")
}
\arguments{
\item{runs}{output file with runs per animal (breed, id, chrom, nSNP, start, end, length) #defaults to detectRUNS.ROHet.csv}
}
\value{
plot of stacked runs by population and by chromosome (pdf files)
}
\description{
PLOT STACKED RUNS
Function to plot stacked runs along the chromosome (signalling presence of large numbers of runs)
Counts on the y-axis, bps on the x-axis: plots run (TRUE) / no run (FALSE)
}
\examples{
#not yet


}

