% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{readPOPCpp}
\alias{readPOPCpp}
\title{Function to return a dataframe of population (POP, ID)}
\usage{
readPOPCpp(genotypeFile)
}
\arguments{
\item{genotypeFile}{genotype (.ped) file location}
}
\value{
a dataframe of POP, ID
}
\description{
This is a core function. Read PED file and returns a data.frame with the first two
columns
}
