% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{rohIslands}
\alias{rohIslands}
\title{Deprecated: use \code{\link{runsIslands}} instead}
\usage{
rohIslands(...)
}
\arguments{
\item{...}{passed to \code{runsIslands}}
}
\description{
Deprecated: use \code{\link{runsIslands}} instead
}
\keyword{internal}
