% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{RUNS.run}
\alias{RUNS.run}
\title{Run functions to detect RUNS (ROHom/ROHet)}
\usage{
RUNS.run(gegevens, mappa, windowSize = 15, drempel = 0.1, minSNP = 3,
  ROHet = TRUE, maxOppositeGenotype = 1, maxMiss = 1)
}
\arguments{
\item{gegevens}{genotype dataset in the Plink .raw format}

\item{mappa}{Plink map file}

\item{drempel}{the threshold of overlapping windows of the same state (homozygous/heterozygous) to call a SNP in a RUN}

\item{minSNP}{minimum n. of SNP in a RUN}

\item{ROHet}{should we look for ROHet or ROHom?}

\item{window}{the size of sliding window}
}
\value{
n. of individuals for which runs have been written out
}
\description{
This is the main function of RUNS and would probably require
some good documentation.
}
\examples{
x <- RUNS.run(gegevens, mappa, windowSize = 20, drempel = 0.1, minSNP = 5, ROHet = TRUE, maxOppositeGenotype = 1, maxMiss = 1)

}

