% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{RUNS.run}
\alias{RUNS.run}
\title{Run functions to detect RUNS (ROHom/ROHet)}
\usage{
RUNS.run(genotype_path, mapfile_path, windowSize = 15, threshold = 0.1,
  minSNP = 3, ROHet = TRUE, maxOppositeGenotype = 1, maxMiss = 1,
  maxGap = 10^6, minLengthBps = 1000, minDensity = 1/10)
}
\arguments{
\item{genotype_path}{genotype (.ped) file location}

\item{mapfile_path}{map file (.map) file location}

\item{windowSize}{the size of sliding window}

\item{threshold}{the threshold of overlapping windows of the same state (homozygous/heterozygous) to call a SNP in a RUN}

\item{minSNP}{minimum n. of SNP in a RUN}

\item{ROHet}{should we look for ROHet or ROHom?}

\item{maxOppositeGenotype}{max n. of homozygous/heterozygous SNP}

\item{maxMiss}{max. n. of missing SNP}

\item{maxGap}{max distance between consecutive SNP in a window to be stil considered a potential run}

\item{minLengthBps}{minimum length of run in bps (defaults to 1000 bps = 1 kbps)}

\item{minDensity}{minimum n. of SNP per kbps (defaults to 0.1 = 1 SNP every 10 kbps)}
}
\value{
a dataframe with RUNs of Homozygosity or Heterozygosity
}
\description{
This is the main function of RUNS and would probably require
some good documentation.
}
\examples{
# getting map and ped paths
genotype_path <- system.file("extdata", "subsetChillingham.ped", package = "detectRUNS")
mapfile_path <- system.file("extdata", "subsetChillingham.map", package = "detectRUNS")
x <- RUNS.run(genotype_path, mapfile_path, windowSize = 20, threshold = 0.1, minSNP = 5,
ROHet = TRUE, maxOppositeGenotype = 1, maxMiss = 1,  minLengthBps = 1000, minDensity = 1/10)

}

