% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{saveROH}
\alias{saveROH}
\title{Deprecated: use \code{\link{saveRUNS}} instead}
\usage{
saveROH(...)
}
\arguments{
\item{...}{passed to \code{saveRUNS}}
}
\description{
Deprecated: use \code{\link{saveRUNS}} instead
}
\keyword{internal}
