% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{schiebeFenster}
\alias{schiebeFenster}
\title{Function to slide a window over a vector (individual's genotypes)}
\usage{
schiebeFenster(data, gaps, fenster, step, ROHet = TRUE,
  maxOppositeGenotype = 1, maxMiss = 1, maxGap)
}
\arguments{
\item{gaps}{vector of differences between consecutive positions (gaps) in bps}

\item{step}{by which (how many SNP) is the window slidden}

\item{ROHet}{shall we detect ROHet or ROHom?}

\item{maxOppositeGenotype}{max n. of homozygous/heterozygous SNP}

\item{maxMiss}{max. n. of missing SNP}

\item{maxGap}{max distance between consecutive SNP in a window to be stil considered a potential run}

\item{x}{vector of 0/1/2 genotypes}

\item{window}{size of window (n. of SNP)}
}
\value{
vector of TRUE/FALSE (whether a window is homozygous or NOT)
}
\description{
This is a core function. The functions to detect RUNS are slidden over the genome
}
\examples{
#not yet


}

