% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run.R
\name{slidingRUNS.run}
\alias{slidingRUNS.run}
\title{Main function to detect RUNS (ROHom/ROHet) using sliding windows (a la Plink)}
\usage{
slidingRUNS.run(
  genotypeFile,
  mapFile,
  windowSize = 15,
  threshold = 0.05,
  minSNP = 3,
  ROHet = FALSE,
  maxOppWindow = 1,
  maxMissWindow = 1,
  maxGap = 10^6,
  minLengthBps = 1000,
  minDensity = 1/1000,
  maxOppRun = NULL,
  maxMissRun = NULL,
  nCores = NULL
)
}
\arguments{
\item{genotypeFile}{genotype (.ped) file path}

\item{mapFile}{map file (.map) file path}

\item{windowSize}{the size of sliding window (number of SNP loci) (default = 15)}

\item{threshold}{the threshold of overlapping windows of the same state
(homozygous/heterozygous) to call a SNP in a RUN (default = 0.05)}

\item{minSNP}{minimum n. of SNP in a RUN (default = 3)}

\item{ROHet}{should we look for ROHet or ROHom? (default = FALSE)}

\item{maxOppWindow}{max n. of homozygous/heterozygous SNP in the
sliding window (default = 1)}

\item{maxMissWindow}{max. n. of missing SNP in the sliding window (default = 1)}

\item{maxGap}{max distance between consecutive SNP to be still considered a
potential run (default = 10^6 bps)}

\item{minLengthBps}{minimum length of run in bps (defaults to 1000 bps = 1 kbps)}

\item{minDensity}{minimum n. of SNP per kbps (defaults to 0.1 = 1 SNP every 10 kbps)}

\item{maxOppRun}{max n. of opposite genotype SNPs in the run (optional)}

\item{maxMissRun}{max n. of missing SNPs in the run (optional)}

\item{nCores}{number of cores to use (deprecated, ignored)}
}
\value{
A dataframe with RUNs of Homozygosity or Heterozygosity.
}
\description{
\strong{Deprecated.}
}
\details{
This function is deprecated. Use \code{\link{scanRUNS}} instead.
}
\examples{
# getting map and ped paths
genotypeFile <- system.file("extdata", "Kijas2016_Sheep_subset.ped", package = "detectRUNS")
mapFile <- system.file("extdata", "Kijas2016_Sheep_subset.map", package = "detectRUNS")
\dontrun{
runs <- slidingRUNS.run(genotypeFile, mapFile, windowSize = 15, threshold = 0.1,
minSNP = 15, ROHet = FALSE,  maxOppWindow = 1, maxMissWindow = 1, maxGap=10^6,
minLengthBps = 100000,  minDensity = 1/10000)
}
runsFile <- system.file("extdata", "Kijas2016_Sheep_subset.sliding.csv", package="detectRUNS")
colClasses <- c(rep("character", 3), rep("numeric", 4))
runs <- read.csv2(runsFile, header = TRUE, stringsAsFactors = FALSE, colClasses = colClasses)

}
