% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{slidingWindow}
\alias{slidingWindow}
\title{Function to slide a window over a vector (individual's genotypes)}
\usage{
slidingWindow(data, gaps, windowSize, step, maxGap, ROHet = TRUE,
  maxOppositeGenotype = 1, maxMiss = 1)
}
\arguments{
\item{data}{vector of 0/1/2 genotypes}

\item{gaps}{vector of differences between consecutive positions (gaps) in bps}

\item{windowSize}{size of window (n. of SNP)}

\item{step}{by which (how many SNP) is the window slidden}

\item{maxGap}{max distance between consecutive SNP in a window to be stil considered a potential run}

\item{ROHet}{shall we detect ROHet or ROHom?}

\item{maxOppositeGenotype}{max n. of homozygous/heterozygous SNP}

\item{maxMiss}{max. n. of missing SNP}
}
\value{
vector of TRUE/FALSE (whether a window is homozygous or NOT)
}
\description{
This is a core function. The functions to detect RUNS are slidden over the genome
}
\examples{
windowSize <- 5
step= 1
maxOppositeGenotype <- 1
maxMiss <- 1
maxGap <- 10^5
ROHet= FALSE
x <- c(0, 0, NA, 0, NA, 0, 0, 0, 1, 1,
       1, 1, 1, 1, 0, 0, 1, 0, 0, 0)
gaps <- c(4374, 8744, 5123, 14229, 5344, 690, 8566, 5853, 2369, 3638,
          4848, 600, 2333, 976, 2466, 2269, 5411, 6021, 4367)
test <- slidingWindow(x, gaps, windowSize, step, maxGap, ROHet, maxOppositeGenotype, maxMiss)
# test is
# $windowStatus
# [1] FALSE FALSE FALSE  TRUE  TRUE FALSE FALSE FALSE FALSE FALSE FALSE FALSE
# FALSE FALSE  TRUE  TRUE

# $oppositeAndMissingGenotypes
#   3   5   9  10  11  12  13  14  17
# "9" "9" "0" "0" "0" "0" "0" "0" "0"

}

