% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{snp_inside_ROH}
\alias{snp_inside_ROH}
\title{Function to count number of times a SNP is in a RUN}
\usage{
snp_inside_ROH(runsFile, mapFile = "karte.map")
}
\arguments{
\item{runsFile}{file with runs (results): output from RUNS.run()}

\item{mapFile}{optional map file (defaults to karte.map: created by RUNS.run())}
}
\value{
TRUE/FALSE if SNP counts are written out or not
}
\description{
Function to count number of times a SNP is in a RUN
}
\examples{
#not yet


}

