% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{snp_inside_ROH}
\alias{snp_inside_ROH}
\title{Function to count number of times a SNP is in a RUN}
\usage{
snp_inside_ROH(runs, mapChrom, popFile = "genotype.raw")
}
\arguments{
\item{runs}{R object (dataframe) with results per chromosome: subsetted output from RUNS.run()}

\item{mapChrom}{R map object with SNP per chromosome}

\item{popFile}{file with two columns POPULATION/ID (defaults to genotype.raw, generated by RUNS.run())}
}
\value{
dataframe with counts per SNP in runs (per population)
}
\description{
Function to count number of times a SNP is in a RUN
}
\examples{
#not yet

}

