% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funktionen.R
\name{snpInRun}
\alias{snpInRun}
\title{Function to return a vector of T/F for whether a SNP is or not in a RUN}
\usage{
snpInRun(RunVector, windowSize, threshold)
}
\arguments{
\item{RunVector}{vector of TRUE/FALSE (is a window homozygous/heterozygous?)}

\item{windowSize}{size of window (n. of SNP)}

\item{threshold}{threshold to call a SNP in a RUN}
}
\value{
vector of TRUE/FALSE (whether a SNP is in a RUN or NOT)
}
\description{
This is a core function. The function to determine whether a SNP is or not in a RUN.
The ratio between homozygous/heterozygous windows and total n. of windows is computed here
}
