% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{snpInsideRunsCpp}
\alias{snpInsideRunsCpp}
\title{Function to count number of times a SNP is in a RUN}
\usage{
snpInsideRunsCpp(runsChrom, mapChrom, genotypeFile)
}
\arguments{
\item{runsChrom}{R object (dataframe) with results per chromosome}

\item{mapChrom}{R map object with SNP per chromosome}

\item{genotypeFile}{genotype (.ped) file location}
}
\value{
dataframe with counts per SNP in runs (per population)
}
\description{
Function to count number of times a SNP is in a RUN
}
